Fix license-compiling
